requirements.txt
asgiref==3.6.0
asttokens==2.0.5
autopep8==1.7.0
backcall==0.2.0
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
Django==4.1.4
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.14.0
entrypoints==0.4
executing==0.8.3
ipykernel==6.13.0
ipython==8.2.0
jedi==0.18.1
jupyter-client==7.2.2
jupyter-core==4.10.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.9.1
Pygments==2.11.2
pyparsing==3.0.8
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7
pywin32==303
pyzmq==22.3.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.2.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
tzdata==2022.2
wcwidth==0.2.5
